# talksy-ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI Specification",
    "description": "OpenApi documentation for Talksy API",
    "termsOfService": "https://cybercodecracker.github.io/portfolio/terms",
    "contact": {
      "name": "Talksy API",
      "url": "https://cybercodecracker.github.io/portfolio/",
      "email": "contact@amouri-coding.com"
    },
    "license": {
      "name": "License",
      "url": "https://cybercodecracker.github.io/portfolio/license"
    },
    "version": "0.1"
  },
  "servers": [{ "url": "http://localhost:8080", "description": "Local ENV" }],
  "security": [{ "BearerAuth": [] }],
...
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegistrationRequest" }
            }
          },
          "required": true
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthenticationRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/confirm-account": {
      "get": {
        "tags": ["Authentication"],
        "operationId": "confirmAccount",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RegistrationRequest": {
        "required": ["email", "firstName", "lastName", "nickname", "password"],
        "type": "object",
        "properties": {
          "firstName": {
            "maxLength": 10,
            "minLength": 3,
            "type": "string",
            "example": "Souhail"
          },
          "lastName": {
            "maxLength": 10,
            "minLength": 3,
            "type": "string",
            "example": "Amouri"
          },
          "nickname": {
            "maxLength": 15,
            "minLength": 3,
            "type": "string",
            "example": "CoolDude"
          },
          "email": { "type": "string", "example": "souhail@gmail.com" },
          "password": {
            "maxLength": 15,
            "minLength": 8,
            "pattern": "^(?=.*[A-Za-z])(?=.*\\d)[A-Za-z\\d]{8,}$",
            "type": "string"
          }
        }
      },
      "RefreshRequest": {
        "type": "object",
        "properties": { "refreshToken": { "type": "string" } }
      },
      "AuthenticationResponse": {
        "type": "object",
        "properties": {
          "access_token": { "type": "string" },
          "refresh_token": { "type": "string" },
          "token_type": { "type": "string" }
        }
      },
      "AuthenticationRequest": {
        "required": ["email", "password"],
        "type": "object",
        "properties": {
          "email": { "type": "string", "example": "souhail@gmail.com" },
          "password": { "type": "string", "example": "password123" }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}